requirements.txt
numpy==1.26.3
pandas==1.5.3
pandas-stubs==1.5.3.230214
pyarrow==14.0.2
matplotlib==3.8.2
seaborn==0.13.1
types-seaborn==0.13.2.20240205
flake8==7.0.0
pylint==3.0.3
autopep8==2.0.4
yapf==0.40.2
black==23.12.1
black-macchiato==1.3.0
tqdm==4.66.1
gitpython==3.1.41
dvc==3.38.1
dvc-gdrive==3.0.1
